+++ R/readConfiguration.R
@@ -415,7 +415,7 @@ checkScenario <- function(scenario = defaultScenario())
   as.boolean.param <- function(x, name)
   {
     x <- as.integer(x)
-    if (is.na (x) || (x != 0 && x != 1)) {
+    if (!is.na (x) && x != 0L && x != 1L) {
       irace.error (quote.param(name), " must be either 0 or 1.")
     }
     as.logical(x)
@@ -696,6 +696,10 @@ checkScenario <- function(scenario = defaultScenario())
                 " cannot be enabled at the same time.")
   }
 
+  if (is.na(scenario$capping)) {
+    scenario$capping <- (scenario$elitist && scenario$maxTime > 0 && scenario$boundMax > 0)
+  }
+
   if (scenario$capping) {
     if (!scenario$elitist) 
       irace.error("When capping == TRUE, elitist must be enabled.")
@@ -850,7 +854,7 @@ printScenario <- function(scenario)
 #'    }
 #'  \item Adaptive capping:
 #'    \describe{
-#'      \item{`capping`}{Enable the use of adaptive capping, a technique designed for minimizing the computation time of configurations. This is only available when \code{elitist} is active. (Default: `0`)}
+#'      \item{`capping`}{Enable the use of adaptive capping, a technique designed for minimizing the computation time of configurations. Capping is enabled by default if \code{elitist} is active, \code{maxTime > 0} and  \code{boundMax > 0}. (Default: `NA`)}
 #'      \item{`cappingType`}{Measure used to obtain the execution bound from the performance of the elite configurations.\itemize{\item median: Median performance of the elite configurations.\item mean: Mean performance of the elite configurations.\item best: Best performance of the elite configurations.\item worst: Worst performance of the elite configurations.} (Default: `"median"`)}
 #'      \item{`boundType`}{Method to calculate the mean performance of elite configurations.\itemize{\item candidate: Mean execution times across the executed instances and the current one.\item instance: Execution time of the current instance.} (Default: `"candidate"`)}
 #'      \item{`boundMax`}{Maximum execution bound for \code{targetRunner}. It must be specified when capping is enabled. (Default: `0`)}